# backend/src/main/java/com/example/blog/auth/dto/LoginRequest.java
package com.example.blog.auth.dto;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public record LoginRequest(
        @NotBlank @Size(max = 64) String username,
        @NotBlank @Size(max = 200) String password
) {
}